+++ elfo-logger/src/config.rs
@@ -7,7 +7,8 @@ pub(crate) struct Config {
     #[serde(default)]
     pub(crate) sink: Sink,
     pub(crate) path: Option<PathBuf>,
-    // TODO: colors
+    #[serde(default)]
+    pub(crate) format: Format,
 }
 
 #[derive(Debug, PartialEq, Deserialize)]
@@ -21,4 +22,13 @@ impl Default for Sink {
     fn default() -> Self {
         Sink::Stdout
     }
+}
+
+#[derive(Debug, Deserialize, Default)]
+pub(crate) struct Format {
+    #[serde(default)]
+    pub(crate) with_location: bool,
+    #[serde(default)]
+    pub(crate) with_module: bool,
+    // TODO: colors
 }