+++ elfo-logger/src/layer/mod.rs
@@ -66,8 +66,8 @@ impl<S: Subscriber> Layer<S> for PrintLayer {
 
         let event = PreparedEvent {
             timestamp: now(),
-            level: *event.metadata().level(),
             trace_id: tls::try_trace_id(),
+            metadata: event.metadata(),
             object: tls::try_meta(),
             span_id: event.parent().or_else(|| current_span.id()).cloned(),
             payload_id,