+++ elfo-logger/src/theme.rs
@@ -12,12 +12,16 @@ pub(crate) trait Theme {
     type TraceId: Formatter<Option<TraceId>>;
     type ObjectMeta: Formatter<Option<Arc<ObjectMeta>>>;
     type Payload: Formatter<str>;
+    type Location: Formatter<(&'static str, u32)>;
+    type Module: Formatter<str>;
 }
 
 pub(crate) struct PlainTheme;
 
 impl Theme for PlainTheme {
     type Level = Level;
+    type Location = Location;
+    type Module = Module;
     type ObjectMeta = EmptyIfNone<Arc<ObjectMeta>>;
     type Payload = Payload;
     type Timestamp = Rfc3339Weak;
@@ -28,6 +32,8 @@ pub(crate) struct ColoredTheme;
 
 impl Theme for ColoredTheme {
     type Level = ColoredLevel;
+    type Location = ColoredLocation;
+    type Module = ColoredModule;
     type ObjectMeta = EmptyIfNone<ColoredByHash<Arc<ObjectMeta>>>;
     type Payload = ColoredPayload;
     type Timestamp = Rfc3339Weak;