+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configurer: reload configs on SIGHUP.
 - logger: get spans back in order to enable filtering by `RUST_LOG`.
 - logger: reopen a log file on SIGHUP and when the config is changed.
+- logger: add `format.with_location` and `format.with_module` options to append `@location` and `@module` fields to logs.
 - `Message::NAME` and `Message::PROTOCOL`.
 
 ### Fixed